Add monitor for Home Assistant (Cloudflare origin)
Because Home Assistant is proxied via Cloudflare, the standard
cert_expiry integration only gives the expiry of the Cloudflare
certificate. The shell_command uses the Cloudflare API to find the
origin for a given hostname and output the certificate expiry date.
The template calls this API daily and creates a sensor that matches
the cert_expiry schema.